+++ public/styles/styles.css
@@ -21,25 +21,19 @@ body {
 }
 
 button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  cursor: pointer;
-  transition: border-color 0.25s;
+  background-color: var(--clr-lighter-blue);
+  border-color: var(--clr-darker-blue);
+  border-radius: 5px;
+  border-width:1px;
+  color: white;
+  font-size: 1rem;
+  padding:0.5rem;
 }
 
 button:hover {
-  border-color: hsl(199, 100%, 20%);
-  color: hsl(199, 100%, 20%);
+  background-color: var(--clr-darker-blue);
 }
 
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
 
 h1 {
   font-size: 3.2em;
@@ -69,21 +63,6 @@ nav a,
   transition: all 0.3s ease;
 }
 
-nav a.active,
-#jump-to a.active {
-  background-color: #04aa6d;
-  color: white;
-}
-
-nav a:hover,
-#jump-to a:hover {
-  background-color: var(--clr-lighter-blue);
-  color: white;
-}
-
-nav a:last-child {
-  text-wrap: nowrap;
-}
 
 #page-info {
   padding-top: 8px;
@@ -97,15 +76,4 @@ nav a:last-child {
   a:hover {
     color: #747bff;
   }
-  button {
-    background-color: #f9f9f9;
-  }
 }
-
-/* .page-summary {
-  padding-top: 20px;
-}
-
-.page-section {
-  padding-left: 2em;
-} */
