+++ src/ui/auth/data/AuthThunk.ts
@@ -4,8 +4,8 @@ import { BaseThunk } from "../../../framework/store/update/BaseThunk.ts";
 import type { AuthRequest } from "../types/AuthRequest.ts";
 import { AuthResponse } from "../types/AuthResponse.ts";
 import { generateApiThunkWithExternalConfig } from "../../../framework/store/update/api/ApiThunkFactory.ts";
-import { getLocalStorageDataIfPresent } from "../../../framework/utils/localStorageUtils.ts";
-import { isAfterNow } from "../../../framework/utils/dateUtils.ts";
+import { getLocalStorageDataIfPresent } from "../../../framework/utils/LocalStorageUtils.ts";
+import { isAfterNow } from "../../../framework/utils/DateUtils.ts";
 import type { AuthReducerError } from "../types/AuthReducerError.ts";
 import { AUTH_TOKEN_KEY } from "../../../utils/params.ts";
 import { getSupabaseClient } from "../SupabaseClient.ts";
